# inventory-be/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Automation and Tooling API",
    "description": "Comprehensive API documentation for the structural inventory platform. Includes Auth, RBAC, Users, Locations, Stocks, Requests, Audit, Notification, and Dashboard modules.",
    "version": "1.1.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000/",
      "description": "Local Development Server"
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "example": "Resource not found."
          }
        }
      },
      "PaginationMeta": {
        "type": "object",
        "properties": {
          "total": { "type": "integer", "example": 100 },
          "page": { "type": "integer", "example": 1 },
          "limit": { "type": "integer", "example": 10 },
          "totalPages": { "type": "integer", "example": 10 }
        }
      },
...
      "Location": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "name": { "type": "string", "example": "Aisle B, Shelf 3" },
          "description": { "type": "string" },
          "floor": { "type": "string", "example": "Ground Floor" },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        }
      },
      "Stock": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "modelNumber": { "type": "string", "example": "SENS-01X" },
          "description": { "type": "string" },
          "brand": { "type": "string", "example": "Siemens" },
          "quantity": { "type": "integer", "example": 50 },
          "uom": { "type": "string", "example": "PCS" },
          "projectType": { "type": "string" },
          "type": { "type": "string", "enum": ["mechanical", "electrical"] },
          "minStockLevel": { "type": "integer", "example": 10 },
          "locationId": { "type": "string", "format": "uuid", "nullable": true },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        }
      },
...
    },
    "parameters": {
      "PageParam": {
        "name": "page", "in": "query",
        "schema": { "type": "integer", "default": 1 }
      },
      "LimitParam": {
        "name": "limit", "in": "query",
        "schema": { "type": "integer", "default": 10 }
      },
...
      "IdPathParam": {
        "name": "id", "in": "path", "required": true,
        "schema": { "type": "string", "format": "uuid" }
      }
    }
  },
  "security": [{ "BearerAuth": [] }],
  "paths": {
...
    "/locations": {
      "get": {
        "summary": "Get all physical locations — paginated (Requires locations:read)",
        "tags": ["Locations"],
        "parameters": [
          { "$ref": "#/components/parameters/PageParam" },
          { "$ref": "#/components/parameters/LimitParam" },
          { "name": "search", "in": "query", "description": "Partial match on name or floor", "schema": { "type": "string" } }
        ],
        "responses": {
          "200": {
            "description": "Paginated list of locations",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": { "type": "array", "items": { "$ref": "#/components/schemas/Location" } },
                    "meta": { "$ref": "#/components/schemas/PaginationMeta" }
                  }
                }
              }
            }
          }
        },
        "operationId": "getLocations"
      },
      "post": {
        "summary": "Create a Location (Requires locations:create)",
        "tags": ["Locations"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name"],
                "properties": {
                  "name": { "type": "string" },
                  "description": { "type": "string" },
                  "floor": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Location created", "content": { "application/json": { "schema": { "type": "object", "properties": { "data": { "$ref": "#/components/schemas/Location" } } } } } },
          "401": { "description": "Unauthorized", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
          "403": { "description": "Forbidden — missing locations:create permission", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        },
        "operationId": "postLocations"
      }
    },
    "/locations/{id}": {
      "get": {
        "summary": "Get Location by ID (Requires locations:read)",
        "tags": ["Locations"],
        "parameters": [{ "$ref": "#/components/parameters/IdPathParam" }],
        "responses": {
          "200": { "description": "Location data", "content": { "application/json": { "schema": { "type": "object", "properties": { "data": { "$ref": "#/components/schemas/Location" } } } } } },
          "404": { "description": "Location not found", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        },
        "operationId": "getLocationsId"
      },
      "patch": {
        "summary": "Partially update a Location (Requires locations:update)",
        "tags": ["Locations"],
        "parameters": [{ "$ref": "#/components/parameters/IdPathParam" }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": { "type": "string" },
                  "description": { "type": "string" },
                  "floor": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Location updated", "content": { "application/json": { "schema": { "type": "object", "properties": { "data": { "$ref": "#/components/schemas/Location" } } } } } },
          "404": { "description": "Location not found", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        },
        "operationId": "patchLocationsId"
      },
      "delete": {
        "summary": "Delete Location (Requires locations:delete)",
        "tags": ["Locations"],
        "parameters": [{ "$ref": "#/components/parameters/IdPathParam" }],
        "responses": {
          "200": { "description": "Deleted location returned", "content": { "application/json": { "schema": { "type": "object", "properties": { "data": { "$ref": "#/components/schemas/Location" } } } } } },
          "404": { "description": "Location not found", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        },
        "operationId": "deleteLocationsId"
      }
    },
    "/locations/{id}/stocks": {
      "get": {
        "summary": "Get stocks at a location (Requires locations:read)",
        "tags": ["Locations"],
        "parameters": [{ "$ref": "#/components/parameters/IdPathParam" }],
        "responses": {
          "200": {
            "description": "Location with its stocks",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "location": { "$ref": "#/components/schemas/Location" },
                        "stocks": { "type": "array", "items": { "$ref": "#/components/schemas/Stock" } }
                      }
                    }
                  }
                }
              }
            }
          },
          "404": { "description": "Location not found", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        },
        "operationId": "getLocationsIdStocks"
      }
    },
    "/public/locations/{id}/stocks": {
      "get": {
        "summary": "Public QR manifest — no auth required",
        "tags": ["Locations"],
        "security": [],
        "parameters": [{ "$ref": "#/components/parameters/IdPathParam" }],
        "responses": {
          "200": {
            "description": "Location manifest",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "location": { "$ref": "#/components/schemas/Location" },
                        "stocks": { "type": "array", "items": { "$ref": "#/components/schemas/Stock" } }
                      }
                    }
                  }
                }
              }
            }
          },
          "404": { "description": "Location not found", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        },
        "operationId": "getPublicLocationsIdStocks"
      }
    },
...
    "/stocks/{id}/locations": {
      "get": {
        "summary": "Get location distribution for all stocks of same modelNumber (Requires stocks:read)",
        "tags": ["Stocks"],
        "parameters": [{ "$ref": "#/components/parameters/IdPathParam" }],
        "responses": {
          "200": {
            "description": "Location distribution array",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "locationId": { "type": "string", "format": "uuid" },
                          "locationName": { "type": "string" },
                          "floor": { "type": "string" },
                          "quantity": { "type": "integer" }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "404": { "description": "Stock not found", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        },
        "operationId": "getStocksIdLocations"
      }
    },
...
  }
}